refactor(account): move /account into Management → My Account tab
Account page now lives as a tab inside Management, so the user has a
single place for everything settings-related (Teams, Users, Models,
Account, Company, API, Billing, Integration). The previous "Coming
soon." placeholder under My Account is replaced with the existing
account view, extracted into a reusable component.

  - new components/management/account-tab.tsx — full content of the
    old /account page (header, tier + permissions, profile fields,
    infrastructure choice, connected providers, knowledge documents).
    Drops the trailing "Back to dashboard" button (the page is now
    inside the tab strip, no need for a backlink) and updates the
    "no providers" copy to point at Management → Integration where
    keys are actually managed.
  - apps/(app)/teams/page.tsx renders <AccountTab /> in the my-account
    PageTabsContent slot.
  - apps/(app)/account/page.tsx becomes a server-side redirect to
    /teams?tab=my-account so old bookmarks / external links don't
    404. The route stays in the app tree to absorb stale traffic.
  - sidebar avatar dropdown "My account" item links straight to
    /teams?tab=my-account.
  - route-config drops the /account entry — the redirect runs before
    the appbar would render anyway, so the config was dead.

No BE changes — same fetchOnboardingProfile call, same data, just
relocated client-side.
