+++ apps/web/src/lib/route-config.ts
@@ -26,12 +26,6 @@ const ROUTE_CONFIGS: Record<string, RouteConfig> = {
     hideNotifications: true,
     appbarType: "aiChat",
   },
-  "/account": {
-    bg: "bg-bg-1",
-    title: "Account",
-    hideSearch: true,
-    hideNotifications: true,
-  },
   "/teams": {
     bg: "bg-bg-1",
     title: "Management",
